+++ lib/ui/notes/note_type.dart
@@ -1,11 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:reading_app/theme/theme.dart';
+
 enum NoteType {
-  content(str: "內容筆記"),
-  experience(str:"經驗"),
-  action(str:"行動"),
-  thought(str:"連結想法");
+  content(str: "內容筆記", colorKey: 'primary'),
+  experience(str: "經驗", colorKey: 'secondary'),
+  action(str: "行動", colorKey: 'primaryContainer'),
+  thought(str: "連結想法", colorKey: 'onPrimaryFixed');
 
   final String str;
-  const NoteType({required this.str});
+  final String colorKey;
 
-} 
+  const NoteType({
+    required this.str,
+    required this.colorKey,
+  });
 
+  Color get color {
+    final colorScheme = MaterialTheme.lightScheme();
+    switch (colorKey) {
+      case 'primary':
+        return colorScheme.primaryFixedDim;
+      case 'secondary':
+        return colorScheme.secondaryFixedDim;
+      case 'primaryContainer':
+        return colorScheme.tertiaryFixedDim;
+      case 'onPrimaryFixed':
+        return colorScheme.surfaceDim;
+      default:
+        return Colors.black; // Fallback color
+    }
+  }
+}
