+++ pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   cupertino_icons: ^1.0.6
   go_router: ^14.2.3
   provider: ^6.1.2
+  intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
